+++ GPTifier/CMakeLists.txt
@@ -82,13 +82,13 @@ add_executable(
   "src/command_chats.cpp"
   "src/command_files.cpp"
   "src/command_fine_tune.cpp"
-  "src/command_models.cpp"
   "src/command_test.cpp"
   "src/configs.cpp"
   "src/datadir.cpp"
   "src/help_messages.cpp"
   "src/interface/command_costs.cpp"
   "src/interface/command_embed.cpp"
+  "src/interface/command_models.cpp"
   "src/interface/command_run.cpp"
   "src/interface/command_short.cpp"
   "src/main.cpp"
