+++ GPTifier/include/interface/command_models.hpp
